requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
python-dotenv==1.0.0
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.5.0
langchain==0.1.1
openai==1.3.0
